Test for float8
